requirements.txt
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0
pandas>=2.3.2
SQLAlchemy>=2.0.43
